+++ Telegram.BotAPI/Parameters/AvailableMethods/CopyMessageParameters.cs
@@ -8,9 +8,9 @@ public sealed class CopyMessageParameters : ApiRequestParameters
 {
     public required ChatIdSource ChatId { get; init; }
 
-    public int? MessageThreadId { get; init; }
+    public long? MessageThreadId { get; init; }
 
-    public int? DirectMessagesTopicId { get; init; }
+    public long? DirectMessagesTopicId { get; init; }
 
     public required ChatIdSource FromChatId { get; init; }
