+++ Telegram.BotAPI/Parameters/AvailableMethods/DeleteForumTopicParameters.cs
@@ -7,5 +7,5 @@ public sealed class DeleteForumTopicParameters : ApiRequestParameters
 {
     public required ChatIdSource ChatId { get; init; }
 
-    public required int MessageThreadId { get; init; }
+    public required long MessageThreadId { get; init; }
 }
