+++ Telegram.BotAPI/Parameters/AvailableMethods/SendMediaGroupParameters.cs
@@ -10,9 +10,9 @@ public sealed class SendMediaGroupParameters : ApiRequestParameters
 
     public required ChatIdSource ChatId { get; init; }
 
-    public int? MessageThreadId { get; init; }
+    public long? MessageThreadId { get; init; }
 
-    public int? DirectMessagesTopicId { get; init; }
+    public long? DirectMessagesTopicId { get; init; }
 
     public required IReadOnlyList<InputMedia> Media { get; init; }
