+++ Telegram.BotAPI/Parameters/AvailableMethods/SendPaidMediaParameters.cs
@@ -10,9 +10,9 @@ public sealed class SendPaidMediaParameters : ApiRequestParameters
 
     public required ChatIdSource ChatId { get; init; }
 
-    public int? MessageThreadId { get; init; }
+    public long? MessageThreadId { get; init; }
 
-    public int? DirectMessagesTopicId { get; init; }
+    public long? DirectMessagesTopicId { get; init; }
 
     public required int StarCount { get; init; }
