+++ Telegram.BotAPI/Parameters/Payments/SendInvoiceParameters.cs
@@ -8,9 +8,9 @@ public sealed class SendInvoiceParameters : ApiRequestParameters
 {
     public required ChatIdSource ChatId { get; init; }
 
-    public int? MessageThreadId { get; init; }
+    public long? MessageThreadId { get; init; }
 
-    public int? DirectMessagesTopicId { get; init; }
+    public long? DirectMessagesTopicId { get; init; }
 
     public required string Title { get; init; }
